tasks

+++ sort.h
@@ -22,5 +22,5 @@ void bubble_sort(int *array, size_t size);
 void insertion_sort_list(listint_t **list);
 void selection_sort(int *array, size_t size);
 void quick_sort(int *array, size_t size);
-size_t partition(int array[], int start, size_t size);
+size_t partition(int array[], size_t start, size_t size);
 #endif
